feat(notifications): web desktop for notifications page
