+++ src/components/organisms/notification-list/notification-list.module.scss
@@ -4,4 +4,5 @@
     font-size: var(--font-size-s2);
     color: var(--color-primary-01);
     padding: 1.25rem 0;
+    cursor: pointer;
 }
