+++ src/pages/notifications/desktop/desktop.module.scss
@@ -1,5 +1,17 @@
-.settings {
+.rightContainer {
   display: flex;
   align-items: center;
   justify-content: space-between;
+  position: sticky;
+  top: 2.5rem;
+
+  .avatar {
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+  }
+
+  .icon {
+    cursor: pointer;
+  }
 }
