--- src/pages/notifications/desktop/desktop.types.ts
@@ -1,5 +1,0 @@
-import { NotificationItemProps } from "../../../components/molecules/notification-item/notification-item.types";
-
-export type NotificationDesktopProps = {
-    list: NotificationItemProps[];
-}
