--- src/pages/notifications/mobile/mobile.service.ts
@@ -1,6 +1,0 @@
-import { get } from "../../../core/http";
-
-
-export async function getNotificationList({ page } = { page: 1 }) {
-    return get(`notifications?page=${page}`).then(({ data }) => data);
-  }
