--- src/pages/notifications/mobile/mobile.types.ts
@@ -1,7 +1,0 @@
-import { Pagination } from "../../../core/types";
-import { Notifications } from "../../../components/organisms/notification-list/notification-list.types";
-
-export type NotificationMobileProps = {
-    list: Pagination<Notifications[]>;
-   
-}
