+++ src/pages/notifications/notifications.shared.tsx
@@ -1,0 +1,34 @@
+import { useState } from 'react';
+import { useMatch } from '@tanstack/react-location';
+import { useSelector } from 'react-redux';
+import { RootState } from 'src/store/store';
+import { getNotificationList } from './notifications.service';
+import { IdentityReq } from 'src/core/types';
+import { Resolver } from './notifications.types';
+
+export const useNotificationsShared = () => {
+  const list = useMatch().ownData as Resolver;
+  const [notificationList, setNotificationList] = useState(list.items);
+  const [page, setPage] = useState(1);
+  const totalCount = list.total_count;
+  const identity = useSelector<RootState, IdentityReq>((state) => {
+    return state.identity.entities.find((identity) => identity.current) as IdentityReq;
+  });
+  const avatarImg = identity?.meta?.avatar || identity?.meta?.image;
+
+  const onShowSeeMore = (length: number): boolean => {
+    if (length < totalCount) {
+      return true;
+    }
+    return false;
+  };
+
+  const onMorePageClick = () => {
+    getNotificationList({ page: page + 1 }).then((resp) => {
+      setPage((v) => v + 1);
+      setNotificationList((list) => [...list, ...resp.items]);
+    });
+  };
+
+  return { notificationList, identity, avatarImg, onMorePageClick, onShowSeeMore };
+};
