+++ src/pages/notifications/notifications.types.ts
@@ -1,0 +1,4 @@
+import { Notifications } from 'src/components/organisms/notification-list/Notification-list.types';
+import { Pagination } from 'src/core/types';
+
+export type Resolver = Pagination<Notifications[]>;
